replace dependents encrypted jobs with regular jobs except code
