+++ app/sidekiq/bgs/submit_form686c_encrypted_job.rb
@@ -16,17 +16,21 @@ module BGS
     sidekiq_retries_exhausted do |msg, error|
       user_uuid, icn, saved_claim_id, encrypted_vet_info = msg['args']
       vet_info = JSON.parse(KmsEncrypted::Box.new.decrypt(encrypted_vet_info))
-      Rails.logger.error('BGS::SubmitForm686cJob failed, retries exhausted!',
+      Rails.logger.error('BGS::SubmitForm686cEncryptedJob failed, retries exhausted!',
                          { user_uuid:, saved_claim_id:, icn:, error: })
-      user ||= BGS::SubmitForm686cEncryptedJob.generate_user_struct(vet_info)
-      CentralMail::SubmitCentralForm686cJob.perform_async(saved_claim_id,
-                                                          KmsEncrypted::Box.new.encrypt(vet_info.to_json),
-                                                          KmsEncrypted::Box.new.encrypt(user.to_h.to_json))
+      if Flipper.enabled?(:dependents_central_submission)
+        user ||= BGS::SubmitForm686cEncryptedJob.generate_user_struct(vet_info)
+        CentralMail::SubmitCentralForm686cJob.perform_async(saved_claim_id,
+                                                            KmsEncrypted::Box.new.encrypt(vet_info.to_json),
+                                                            KmsEncrypted::Box.new.encrypt(user.to_h.to_json))
+      else
+        DependentsApplicationFailureMailer.build(user).deliver_now if user&.email.present? # rubocop:disable Style/IfInsideElse
+      end
     end
 
     def perform(user_uuid, icn, saved_claim_id, encrypted_vet_info)
       @vet_info = JSON.parse(KmsEncrypted::Box.new.decrypt(encrypted_vet_info))
-      Rails.logger.info('BGS::SubmitForm686cJob running!', { user_uuid:, saved_claim_id:, icn: })
+      Rails.logger.info('BGS::SubmitForm686cEncryptedJob running!', { user_uuid:, saved_claim_id:, icn: })
 
       @user = BGS::SubmitForm686cEncryptedJob.generate_user_struct(@vet_info)
       @user_uuid = user_uuid
@@ -40,13 +44,13 @@ module BGS
       BGS::Form686c.new(user, claim).submit(claim_data)
 
       # If Form 686c job succeeds, then enqueue 674 job.
-      BGS::SubmitForm674EncryptedJob.perform_async(user_uuid, icn, saved_claim_id, encrypted_vet_info, KmsEncrypted::Box.new.encrypt(user.to_h.to_json)) if claim.submittable_674? # rubocop:disable Layout/LineLength
+      BGS::SubmitForm674Job.perform_async(user_uuid, icn, saved_claim_id, encrypted_vet_info, KmsEncrypted::Box.new.encrypt(user.to_h.to_json)) if claim.submittable_674? # rubocop:disable Layout/LineLength
 
       send_confirmation_email
       in_progress_form&.destroy
-      Rails.logger.info('BGS::SubmitForm686cJob succeeded!', { user_uuid:, saved_claim_id:, icn: })
+      Rails.logger.info('BGS::SubmitForm686cEncryptedJob succeeded!', { user_uuid:, saved_claim_id:, icn: })
     rescue => e
-      Rails.logger.warn('BGS::SubmitForm686cJob received error, retrying...',
+      Rails.logger.warn('BGS::SubmitForm686cEncryptedJob received error, retrying...',
                         { user_uuid:, saved_claim_id:, icn:, error: e.message })
       log_message_to_sentry(e, :warning, {}, { team: 'vfs-ebenefits' })
       salvage_save_in_progress_form(FORM_ID, user_uuid, @in_progress_copy) if @in_progress_copy.present?
