+++ config/features.yml
@@ -330,10 +330,6 @@ features:
     actor_type: user
     description: Manage dependent removal from view dependent page
     enable_in_development: true
-  dependents_encrypt_jobs:
-    actor_type: user
-    description: Manage whether Sidekiq Jobs used for Dependents Claims will encrypt arguments considered potential PII.
-    enable_in_development: true
   dependents_central_submission_lighthouse:
     actor_type: user
     description: Manage whether central processing submission uses Lighthouse API
